pyproject.toml
[project]
name = "llama-tokenizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["transformers>=4.21.0", "torch>=1.12.0", "python-dotenv>=1.0.0"]
